refactor: split session review panel
